dump: Try to fix date reordering issues
Keep frames in a queue so we can reorder and send them according to date
received.
This still has some issues because timestamps are recorded on the host
when a USB URB is received. Since each USB message contains multiple E1
frames in the same direction the timestamps don't necessarily match up.

To work around that there is now an option -a to adjust the timestamp of
a given endpoint by 1ms.